requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
attrs==20.2.0
Django==3.1.1
django-filter==2.4.0
django-graphql-auth==0.3.12
django-graphql-jwt==0.3.0
Faker==0.9.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.0.1
mixer==6.1.3
packaging==20.4
pluggy==0.13.1
promise==2.3
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.0
pytest-django==3.10.0
python-dateutil==2.8.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
Unidecode==1.1.1
